Cleanup Yolov8x and Yolov10x demo_utils
