+++ models/demos/yolov8x/demo/demo.py
@@ -11,14 +11,15 @@ import torch
 from loguru import logger
 
 import ttnn
-from models.demos.utils.common_demo_utils import LoadImages, load_coco_class_names
-from models.demos.yolov8x.common import YOLOV8X_L1_SMALL_SIZE, load_torch_model
-from models.demos.yolov8x.demo.demo_utils import (
+from models.demos.utils.common_demo_utils import (
+    LoadImages,
     get_mesh_mappers,
-    postprocess,
+    load_coco_class_names,
     preprocess,
     save_yolo_predictions_by_model,
 )
+from models.demos.yolov8x.common import YOLOV8X_L1_SMALL_SIZE, load_torch_model
+from models.demos.yolov8x.demo.demo_utils import postprocess
 from models.demos.yolov8x.reference import yolov8x
 from models.demos.yolov8x.runner.performant_runner import YOLOv8xPerformantRunner
 from models.utility_functions import disable_persistent_kernel_cache
